+++ cypress/integration/quickmark/instance-quickmark.spec.js
@@ -6,14 +6,15 @@ import InventoryInstance from '../../support/fragments/inventory/inventoryInstan
 import QuickMarcEditor from '../../support/fragments/quickMarcEditor';
 import InventoryViewSource from '../../support/fragments/inventory/inventoryViewSource';
 import InventoryInstanceEdit from '../../support/fragments/inventory/InventoryInstanceEdit';
+import { testType, feature } from '../../support/utils/tagTools';
 
-describe('Manage records ', () => {
+describe('Manage inventory Bib records with quickMarc editor', () => {
   beforeEach(() => {
     // TODO: add support of special permissions in special account
     cy.login(Cypress.env('diku_login'), Cypress.env('diku_password'));
     cy.visit(TopMenu.inventoryPath);
   });
-  it('C10950 Edit and save a MARC record in quickMARC', () => {
+  it('C10950 Edit and save a MARC record in quickMARC', { tags: [testType.smoke, feature.quickMarcEditor] }, () => {
     InventoryActions.import();
     InventoryInstance.goToEditMARCBiblRecord();
 
@@ -33,7 +34,7 @@ describe('Manage records ', () => {
     // InventoryViewSource.notContains('948\t   \t‡h NO HOLDINGS IN PAOLF - 43 OTHER HOLDINGS ');
   });
 
-  it('C10924 Add a field to a record using quickMARC', () => {
+  it('C10924 Add a field to a record using quickMARC', { tags: [testType.smoke, feature.quickMarcEditor] }, () => {
     InventoryActions.import();
     InventoryInstance.goToEditMARCBiblRecord();
     QuickMarcEditor.addRow();
@@ -50,7 +51,7 @@ describe('Manage records ', () => {
     QuickMarcEditor.checkContent();
   });
 
-  it('C10928 Delete a field(s) from a record in quickMARC', () => {
+  it('C10928 Delete a field(s) from a record in quickMARC', { tags: [testType.smoke, feature.quickMarcEditor] }, () => {
     InventoryActions.import();
     InventoryInstance.goToEditMARCBiblRecord();
     // TODO: add return value to validate the result by value too. As minimum add catching from response. The best way - through interactors
@@ -63,13 +64,13 @@ describe('Manage records ', () => {
     // InventoryViewSource.notContains('948\t   \t‡h NO HOLDINGS IN PAOLF - 43 OTHER HOLDINGS ');
   });
 
-  it('C10957 Attempt to delete a required field', () => {
+  it('C10957 Attempt to delete a required field', { tags: [testType.smoke, feature.quickMarcEditor] }, () => {
     InventoryActions.import();
     InventoryInstance.goToEditMARCBiblRecord();
     QuickMarcEditor.checkRequiredFields();
   });
 
-  it('C10951 Add a 5XX field to a marc record in quickMARC', () => {
+  it('C10951 Add a 5XX field to a marc record in quickMARC', { tags: [testType.smoke, feature.quickMarcEditor] }, () => {
     InventoryActions.import();
     InventoryInstance.overlaySourceBibRecord();
 
@@ -96,7 +97,7 @@ describe('Manage records ', () => {
     InventoryInstance.checkInstanceNotes(testRecord.tagMeaning, testRecord.content);
   });
 
-  it('C345388 Derive a MARC bib record', () => {
+  it('C345388 Derive a MARC bib record', { tags: [testType.smoke, feature.quickMarcEditor] }, () => {
     InventoryActions.import();
     // TODO: check the issue with reading in new version of interactors
     InventoryInstance.getAssignedHRID()