+++ cypress/support/fragments/inventory/newInventoryInstance.js
@@ -1,16 +1,12 @@
-import { MultiColumnList, HTML, including, Button } from '../../../../interactors';
-import { getLongDelay } from '../../utils/cypressTools';
+import { MultiColumnList, HTML, including, Button, Section } from '../../../../interactors';
 
-const rootCss = 'section[id=pane-instancedetails]';
-// TODO: rewrite to interactor with filter by id
-const actionsCss = `${rootCss} button[class*=actionMenuToggle]`;
+const actionsButton = Section({ id:'pane-instancedetails' }).find(Button('Actions'));
 const identifiers = MultiColumnList({ id:'list-identifiers' });
 const editMARCBibRecordButton = Button({ id:'edit-instance-marc' });
 const viewSourceButton = Button({ id:'clickable-view-source' });
 
 
 export default {
-  // TODO: hardcoded count related with interactors getters issue. Redesign to cy.then(QuickMarkEditor().rowsCount()).then(rowsCount => {...}
   validOCLC : { id:'176116217',
     getLastRowNumber:() => 31 },
   checkExpectedOCLCPresence: (OCLCNumber) => {
@@ -18,12 +14,12 @@ export default {
   },
 
   goToEditMARCBiblRecord:() => {
-    cy.get(actionsCss, getLongDelay()).click();
+    cy.do(actionsButton.click());
     cy.do(editMARCBibRecordButton.click());
   },
 
   viewSource: () => {
-    cy.get(actionsCss, getLongDelay()).click();
+    cy.do(actionsButton.click());
     cy.do(viewSourceButton.click());
   }
 };